Another large rework of the model layer.
Iterate on the `refHelpers` idea to create an actual object store.
Everything is now an object with synchronized configuration and run-time
data, references to other objects and garbage collection.

No longer need a private data namespace for plugins. Plugins should
create a root object which already serves as a private namespace.